% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{beta.dev}
\alias{beta.dev}
\title{compute beta-diversity from raw OTU data}
\usage{
beta.dev(OTU_raw, meth)
}
\arguments{
\item{OTU_raw:}{The raw OTU data}

\item{meth:}{The string indicates the method of beta-div computation}
}
\value{
The distance matrix of beta-diversity
}
\description{
compute beta-diversity from raw OTU data
}
