% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pairwise.dist}
\alias{pairwise.dist}
\title{compute beta-diversity from raw OTU data}
\usage{
pairwise.dist(df_cts, meth)
}
\arguments{
\item{df_cts:}{The dataframe with only continuous variables}

\item{meth:}{The string indicates the method of pairwise distance computation}
}
\value{
A list with pairwise distance for each continuous variable
}
\description{
compute beta-diversity from raw OTU data
}
